+++ Disciple/Changes.md
@@ -2,12 +2,12 @@
 
 ## v1.5 (in development)
 ### Breaking Changes
-- Introduced two new constructor parameters to **Gateway**
+- Introduced three new constructor parameters to **Gateway**
   - These should cause only minor errors, if any
 ### New Features
 - Added **RequestRateLimiter** class that limits the number of simultaneous or consecutive actions
 - Added support to SSL certificate configuration in **Gateway**
-  - Pass the `keyStoreSettings` constructor property to use it
+  - Pass the `keyStoreSettings` and/or `trustStoreSettings` constructor property to use it
 - **Gateway** now supports http client customization
 
 ## v1.4.3 - 04.11.2021